% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr2007_03_01-data.R
\docType{data}
\name{cr2007_03_01}
\alias{cr2007_03_01}
\title{Text of the Congressional Record from 3/1/2007}
\format{
A data frame with 154 rows and 4 variables:
\describe{
\item{\code{date}}{date Record date}
\item{\code{congress}}{double congress}
\item{\code{speaker}}{character speaker}
\item{\code{header}}{character header under which the text appears}
\item{\code{url}}{character url to the section of the Record where the text appears}
}
}
\usage{
cr2007_03_01
}
\description{
Text of the Congressional Record from 3/1/2007, scraped and parsed using methods described \href{https://github.com/judgelord/cr}{here}.
}
\details{
A dataframe with parsed text from the U.S. Congressional Record for 2007-03=-01
}
\keyword{datasets}
