% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMemberName.R
\name{extractMemberName}
\alias{extractMemberName}
\title{Extract congress member names from text}
\usage{
extractMemberName(
  data,
  col_name,
  members = legislators::members,
  typos = legislators::typos,
  congress,
  chamber = NULL,
  state = NULL,
  fix_ocr = TRUE,
  verbose = TRUE,
  cl = NULL
)
}
\arguments{
\item{data}{a data frame with a variable (specified in \code{col_name}) containing the text from which members of congress are to be extracted. Can also be a character vector containing the text, which will be converted to a data frame.}

\item{col_name}{\code{character}; when \code{data} is a data frame, the name of the variable containing congress members' names. When \code{data} is a character vector, the name given to the variable containing the original text in the output (if unspecified, \code{"speaker"} will be used).}

\item{members}{a regex table containing variations of congress member names. By default, the \link{members} dataset accompanying the package is used. This table must have the following columns:}

\item{typos}{a dataset from which to extract typos. By default, the \link{typos} dataset accompanying the package is used. This table must have the following columns:}

\item{congress}{the name of a variable in \code{data} containing the congress for each row, or a vector of congress numbers for each row. If a single value is supplied, it will be applied to all rows. The argument is required.}

\item{chamber}{the name of a variable in \code{data} containing the chamber for each row, or a character vector containing the chamber for each row. If a single value is supplied, it will be applied to all rows. Allowable values include the values in \code{members}, which, by default, are \code{"Senate"}, \code{"House"}, and \code{"President"}. \code{NA} values are allowed and will not be incorporated into the match. This argument is optional. See Details.}

\item{state}{the name of a variable in \code{data} containing the state for each row, or a character vector containing the state for each row. If a single value is supplied, it will be applied to all rows. Allowable values include the two-letter abbreviations for each state (e.g., \code{"MA"}) as well as \code{"USA"} for rows corresponding to presidents, or the full names of the states (e.g., \code{"Massachusetts"}). \code{NA} values are allowed and will not be incorporated into the match. This argument is optional. See Details.}

\item{fix_ocr}{\code{logical}; whether to fix OCR errors. Passed to \code{\link[=fix_typos]{fix_typos()}}.}

\item{verbose}{\code{logical}; whether to display information about the process of extracting names, including progress bars.}

\item{cl}{a cluster object created by \code{parallel::makeCluster()}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations. Passed to \code{\link[pbapply:pbapply]{pbapply::pblapply()}}.}
}
\value{
A tibble (data frame) with a row for each match containing the following variables:
\item{data_id}{the row number in \code{data} corresponding to the given match}
\item{icpsr}{the ICPSR ID associated with the matched member}
\item{bioname}{the ICSPR-assigned name of the matched member}
\item{speaker}{the original value used to find matches after processing}
\item{congress}{the congress associated with the matched member}
\item{chamber}{the chamber associated with the matched member}
\item{state_abbrev}{the state (abbreviated) associated with the matched member}
\item{district_code}{the district code associated with the matched member}

In addition, all other variable in \code{data}, including that named in \code{col_name}, will be included in the output.
}
\description{
\code{extractMemberName()} uses a regular expressions lookup to extract names of congress members from supplied text.
}
\details{
\code{extractMemberName()} processes the variable named in \code{col_name} containing the text from which congress members' names are to be extracted. First, it passes the variable to \code{\link[=fix_typos]{fix_typos()}} to apply some heuristic processing, and, if \code{typos} is supplied, it fixes any found typos. Finally, it performs a regular expressions lookup to match congress members listed in \code{members} to the text. For each member in \code{members}, a regular expression match is performed to determine whether the text in the given row contains that member. This is done one congress at a time.

When \code{chamber} or \code{state} are specified, the lookup for each member is restricted to the chamber or state of the member in \code{members}, respectively, which can increase speed and avoid duplicates (e.g., members of a given congress who appear in more than one chamber). These arguments are optional; when not supplied, only the variable named in \code{col_name} and congress are used to identify members. It is possible for a single row in \code{data} to contain multiple members belonging to different states or chambers. In these cases, \code{chambers} or \code{state} can be set to \code{NA} for those rows to avoid using them in their lookup.
}
