% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_typos.R
\name{fix_typos}
\alias{fix_typos}
\title{Fix common typos}
\usage{
fix_typos(x, typos = legislators::typos, fix_ocr = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{a character vector containing the names of congress members.}

\item{typos}{a matrix or data frame containing common typos and their corrections. Should be a 2-column matrix where the first column contains regular expressions corresponding to typos and the second column contain the values that should replace them. These are passed to the \code{pattern} and \code{replacement} arguments of \code{\link[stringr:str_replace]{stringr::str_replace_all()}}. By default, the \link{typos} dataset that accompanies the package is used. If \code{NULL}, no typos will be fixed.}

\item{fix_ocr}{\code{logical}; whether to fix OCR errors.}

\item{verbose}{\code{logical}; whether to display information about the process of fixing typos, including a progress bar.}
}
\value{
A character vector with typos replaced by their corrections.
}
\description{
\code{fix_typos()} standardizes a character vector for use in \code{\link[=extractMemberName]{extractMemberName()}} by applying some heuristic processing, setting all text to lowercase and, optionally, removing common typos if a list of such is supplied.
}
\details{
DETAILS #Explain processing
}
\seealso{
\code{\link[=extractMemberName]{extractMemberName()}}, \code{\link[stringr:str_replace]{stringr::str_replace_all()}}
}
