% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators-package.R
\docType{package}
\name{legislators-package}
\alias{legislators}
\alias{legislators-package}
\title{legislators: Find Legislator Names in Messy Text}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A suite of functions to detect U.S. legislator names in messy text with typos and inconsistent name formats. Example data from the Congressional Record and congressional letters to the Federal Energy Regulatory Commission are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://judgelord.github.io/legislators/}
  \item \url{https://github.com/judgelord/legislators}
  \item Report bugs at \url{https://github.com/judgelord/legislators/issues}
}

}
\author{
\strong{Maintainer}: Devin Judge-Lord \email{judgelor@umich.edu} (\href{https://orcid.org/0000-0002-4384-4380}{ORCID})

Authors:
\itemize{
  \item Noah Greifer
  \item Maggie Nead
}

}
\keyword{internal}
