% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typos-data.R
\docType{data}
\name{typos}
\alias{typos}
\title{Common typos}
\format{
A data frame with 480 rows and 2 variables:
\describe{
\item{\code{typos}}{character COLUMN_DESCRIPTION}
\item{\code{correct}}{character COLUMN_DESCRIPTION}
}
}
\usage{
typos
}
\description{
A dataset contain common typos. Used in the \code{typos} argument of \code{\link[=fix_typos]{fix_typos()}} and \code{\link[=extractMemberName]{extractMemberName()}}.
}
\details{
DETAILS
}
\keyword{datasets}
