% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr2007_03_01-data.R
\docType{data}
\name{cr2007_03_01}
\alias{cr2007_03_01}
\title{Text of the Congressional Record from 3/1/2007}
\format{
A data frame with 154 rows and 5 variables:
\describe{
\item{\code{date}}{double COLUMN_DESCRIPTION}
\item{\code{speaker}}{character COLUMN_DESCRIPTION}
\item{\code{header}}{character COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{url_txt}}{character COLUMN_DESCRIPTION}
}
}
\usage{
cr2007_03_01
}
\description{
Text of the Congressional Record from 3/1/2007, scraped and parsed using methods described \href{https://github.com/judgelord/cr}{here}.
}
\details{
DETAILS
}
\keyword{datasets}
