% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members-data.R
\docType{data}
\name{members}
\alias{members}
\title{Names of members of the US Congress}
\format{
A data frame with 49938 rows and 9 variables:
\describe{
\item{\code{congress}}{double COLUMN_DESCRIPTION}
\item{\code{chamber}}{character COLUMN_DESCRIPTION}
\item{\code{bioname}}{character COLUMN_DESCRIPTION}
\item{\code{pattern}}{character COLUMN_DESCRIPTION}
\item{\code{icpsr}}{double COLUMN_DESCRIPTION}
\item{\code{state_abbrev}}{character COLUMN_DESCRIPTION}
\item{\code{district_code}}{double COLUMN_DESCRIPTION}
\item{\code{first_name}}{character COLUMN_DESCRIPTION}
\item{\code{last_name}}{character COLUMN_DESCRIPTION}
}
}
\usage{
members
}
\description{
A dataset of the names of all members of each United States Congress, including presidents. The \code{pattern} column serves as a lookup table to match members of Congress to text supplied to \code{\link[=extractMemberName]{extractMemberName()}}.
}
\details{
DETAILS
}
\keyword{datasets}
